classes with virtual function must have virtual destructors.
The warning is not "bogus" and not only gcc warns.
clang will warn too, but defines __GNUC__ for gcc compatibility.
msvc issues "warning C4265: 'boost::serialization::<foo>' :
class has virtual functions, but destructor is not virtual"
The msvc warning could be suppressed, but it is much cleaner to simply
make the destructors virtual as they are with gcc and clang.