use std::addressof (or boost::addressof) to make serialization work with class with unary operator&()
I have a class that overloads the unary operator&() the only way I was able to make it work was to modify the library. 
There may be a non intrusive way by creating a proxy return of operator& that can be casted to the right pointer or to the other use but requires a lot of hacking an unnecessary performance hit.